requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-appconf==1.0.6
django-imagekit==5.0.0
django-jazzmin==2.6.0
django-jinja==2.11.0
django-phonenumber-field==7.2.0
django-translations==1.3.0
easy-thumbnails==2.8.5
exception==0.1.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.3
phonenumbers==8.13.26
pilkit==3.0
Pillow==10.1.0
requests==2.31.0
six==1.16.0
sorl-thumbnail==12.3
sqlparse==0.4.4
urllib3==1.25
